requirements.txt
# FastAPI 및 기본 웹 프레임워크
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# 데이터베이스
sqlalchemy==2.0.23
sqlmodel==0.0.10
asyncpg==0.29.0  # PostgreSQL 비동기 드라이버
greenlet==3.0.3  # SQLAlchemy 비동기 지원
alembic==1.13.1

# 인증 및 보안
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 개발 도구
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# 로깅 및 모니터링
python-dotenv==1.0.0
